Update overview layout to fixed table details
In order to have control over what details should render, pick
out the specific ones instead of looping through all of them.

Remove border on header and instead have a slightly different
color of overview in order for it to not look too barren. Also
setting a smaller max widht on the content helps in cases where
there are few details.