Adapt logic to data format and update serach results
Updated the logic to work with the data structure, and with it
moving logic higher up in the component tree. The Product
container contains most of the logic now, and components just
calcs breadkpoints.

Updated the search results so that it uses the same component
as the dropdown, but toggling with props.